Function envmnt::get_or

pub fn get_or<K: AsRef<OsStr>>(key: K, default_value: &str) -> String

Returns the environment variable value or if is not defined, the default value will be returned.

Arguments

key - The environment variable name

default_value - In case the environment variable is not defined, this value will be returned.

Example

fn main() {
    envmnt::set("MY_ENV_VAR", "SOME VALUE");
    assert!(envmnt::is_equal("MY_ENV_VAR", "SOME VALUE"));

    let mut value = envmnt::get_or("MY_ENV_VAR", "DEFAULT_VALUE");
    assert_eq!(value, "SOME VALUE");

    value = envmnt::get_or("ANOTHER_ENV_VAR", "DEFAULT_VALUE");
    assert_eq!(value, "DEFAULT_VALUE");
}